use unsigned long to hold irq state

+++ fs/xfs/xfs_trans_priv.h
@@ -64,7 +64,7 @@ void			xfs_trans_update_ail(struct xfs_mount *,
 				     struct xfs_log_item *, xfs_lsn_t,
 				     unsigned long);
 void			xfs_trans_delete_ail(struct xfs_mount *,
-				     struct xfs_log_item *, int);
+				     struct xfs_log_item *, unsigned long);
 struct xfs_log_item	*xfs_trans_first_ail(struct xfs_mount *, int *);
 struct xfs_log_item	*xfs_trans_next_ail(struct xfs_mount *,
 				     struct xfs_log_item *, int *, int *);
